Code Challenge
Challenge 1: Make the first Rating component a read-only component, so that it cannot be changed by swiping. 
Challenge 2: Both Ratings use stars, which are the default if you do not specify otherwise. Turn them both into rockets. 
Challenge 3: Make it possible to set your rating (the second Rating component) with a fraction of one decimal point, e.g. 4.1 or 2.7.
Challenge 4: If you do not specify a number for the max rating, then it is set by default to 5. Set both Ratings to a max of 10 rockets. 
Challenge 5: Show the second Rating's numeric value in real-time, e.g. Rating: 5.6/10
Challenge 6: Add an onPress prop to the Button component so that when the button is clicked, the setter function of setSubmittedMsg() is called to update the submittedMsg state value to the string 'Submitted Rating: ' followed by the current rating. See the screenshot below for an example. 
Challenge 7: View the Layout Props documentation for React Native, then modify the style prop for each component using the available layout props. There is no single right answer for this challenge; the motive is to get you to experiment with the layout props and see for yourself how they work. Try out different props and values to see for yourself what they do. You can keep the layout props that are already there and add others, or change them entirely. You can choose to use inline styles or use the styles StyleSheet at the bottom.
Challenge 8: Along with the link to your saved Snack, write at least 3 sentences or more about what layout props you tried in Challenge 7, and what you learned about how they work. 
In the column on the right, you have the option to choose the iOS, Android, or Web view for your Snack. You can use any of the views you prefer for this challenge, though the Web view will typically be the fastest.